refactor: HomePage와 PlaceMapPage 폴더 구조 변경, 파일 이름이나 경로와 안 맞는 import문 수정
