+++ src/page/PlaceMapPage.jsx/DetailPage/PhotoGallery.jsx
